+++ tests/test_import_avro.py
@@ -23,7 +23,7 @@ def test_cli():
 
 
 def test_import_avro_schema():
-    result = DataContract().import_from_source("avro", "fixtures/avro/data/orders.avsc")
+    result = DataContract.import_from_source("avro", "fixtures/avro/data/orders.avsc")
 
     expected = """
 dataContractSpecification: 1.2.1
@@ -120,7 +120,7 @@ models:
 
 
 def test_import_avro_arrays_of_records_and_nested_arrays():
-    result = DataContract().import_from_source("avro", "fixtures/avro/data/arrays.avsc")
+    result = DataContract.import_from_source("avro", "fixtures/avro/data/arrays.avsc")
 
     expected = """
 dataContractSpecification: 1.2.1
@@ -171,7 +171,7 @@ models:
 
 
 def test_import_avro_nested_records():
-    result = DataContract().import_from_source("avro", "fixtures/avro/data/nested.avsc")
+    result = DataContract.import_from_source("avro", "fixtures/avro/data/nested.avsc")
 
     expected = """
 dataContractSpecification: 1.2.1
@@ -200,7 +200,7 @@ models:
 
 
 def test_import_avro_nested_records_with_arrays():
-    result = DataContract().import_from_source("avro", "fixtures/avro/data/nested_with_arrays.avsc")
+    result = DataContract.import_from_source("avro", "fixtures/avro/data/nested_with_arrays.avsc")
 
     expected = """
 dataContractSpecification: 1.2.1
@@ -266,7 +266,7 @@ models:
 
 
 def test_import_avro_logical_types():
-    result = DataContract().import_from_source("avro", "fixtures/avro/data/logical_types.avsc")
+    result = DataContract.import_from_source("avro", "fixtures/avro/data/logical_types.avsc")
 
     expected = """
 dataContractSpecification: 1.2.1
@@ -317,7 +317,7 @@ models:
 
 
 def test_import_avro_optional_enum():
-    result = DataContract().import_from_source("avro", "fixtures/avro/data/optional_enum.avsc")
+    result = DataContract.import_from_source("avro", "fixtures/avro/data/optional_enum.avsc")
 
     expected = """
 dataContractSpecification: 1.2.1