+++ tests/test_import_unity_file.py
@@ -26,7 +26,7 @@ def test_cli():
 
 def test_import_unity():
     print("running test_import_unity")
-    result = DataContract().import_from_source("unity", "fixtures/databricks-unity/import/unity_table_schema.json")
+    result = DataContract.import_from_source("unity", "fixtures/databricks-unity/import/unity_table_schema.json")
 
     with open("fixtures/databricks-unity/import/datacontract.yaml") as file:
         expected = file.read()
@@ -56,7 +56,7 @@ def test_cli_complex_types():
 @pytest.mark.skip(reason="Complex types are not perfectly supported for the unity catalog import")
 def test_import_unity_complex_types():
     print("running test_import_unity_complex_types")
-    result = DataContract().import_from_source(
+    result = DataContract.import_from_source(
         "unity", "fixtures/databricks-unity/import/unity_table_schema_complex_types.json"
     )
 
@@ -70,7 +70,7 @@ def test_import_unity_complex_types():
 
 def test_import_unity_with_owner_and_id():
     print("running test_import_unity_with_owner_and_id")
-    result = DataContract().import_from_source(
+    result = DataContract.import_from_source(
         "unity", "fixtures/databricks-unity/import/unity_table_schema.json", owner="sales-team", id="orders-v1"
     )